requirements.txt
solana>=0.36.9
solders>=0.26.0
aiohttp>=3.9
websockets>=12.0
python-dotenv>=1.0
